refactor. user 엔티티 생성 시 default archive 같이 생성

+++ src/main/java/junki/fishkeepingback/domain/image/uploader/S3Uploader.java
@@ -33,6 +33,7 @@ public class S3Uploader {
 
     @Transactional
     public void delete(String fileName) {
+        if (fileName == null) return;
         try {
             log.info("Deleting file {}", fileName);
             amazonS3.deleteObject(bucket, fileName);
